requirements.txt
flask
Pillow
nose
